+++ Semplificata/FatturaElettronicaBody/Allegati/Allegati.cs
