+++ Semplificata/FatturaElettronicaBody/DatiBeniServizi/DatiBeniServizi.cs
