+++ Semplificata/FatturaElettronicaBody/DatiBeniServizi/DatiIVA.cs
