+++ Semplificata/FatturaElettronicaBody/DatiGenerali/DatiFatturaRettificata.cs
