+++ Semplificata/FatturaElettronicaBody/DatiGenerali/DatiGenerali.cs
