+++ Semplificata/FatturaElettronicaBody/DatiGenerali/DatiGeneraliDocumento.cs
