+++ Semplificata/FatturaElettronicaBody/FatturaElettronicaBody.cs
