+++ Semplificata/FatturaElettronicaHeader/CedentePrestatore/CedentePrestatore.cs
