+++ Semplificata/FatturaElettronicaHeader/CedentePrestatore/IscrizioneRea.cs
