+++ Semplificata/FatturaElettronicaHeader/CedentePrestatore/RappresentanteFiscale.cs
