+++ Semplificata/FatturaElettronicaHeader/CessionarioCommittente/AltriDatiIdentificativi.cs
