+++ Semplificata/FatturaElettronicaHeader/CessionarioCommittente/CessionarioCommittente.cs
