+++ Semplificata/FatturaElettronicaHeader/CessionarioCommittente/IdentificativiFiscali.cs
