+++ Semplificata/FatturaElettronicaHeader/CessionarioCommittente/RappresentanteFiscaleCessionarioCommittente.cs
