+++ Semplificata/FatturaElettronicaHeader/CessionarioCommittente/SedeCessionarioCommittente.cs
