+++ Semplificata/FatturaElettronicaHeader/DatiTrasmissione/DatiTrasmissione.cs
