+++ Semplificata/FatturaElettronicaHeader/DatiTrasmissione/IdTrasmittente.cs
