+++ Semplificata/FatturaElettronicaHeader/FatturaElettronicaHeader.cs
